% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{add_rating_param_bits}
\alias{add_rating_param_bits}
\title{Add log likelihood columns to a rating_params data frame}
\usage{
add_rating_param_bits(rating_params)
}
\arguments{
\item{rating_params}{A data frame with columns subject_id, rating, rater_id,
and at least parameters t, a, and p for each row}
}
\value{
The original data frame with columns C0 and C1 added
}
\description{
Add log likelihood columns to a rating_params data frame
}
\details{
This function assumes the (t_i)-a_j-p_j model, whereby raters can
have different opinions about the same subject's probability of Class 1.
DEPRECATED
}
