% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{ap_log_likelihood}
\alias{ap_log_likelihood}
\title{Log likelihood for a,p and confusion matrix}
\usage{
ap_log_likelihood(params, confusion)
}
\arguments{
\item{params}{a vector of parameters a and p}

\item{confusion}{a data frame with columns TP, FP, TN, FN}
}
\value{
the log likelihood of the confusion matrix given the parameters
}
\description{
Log likelihood for a,p and confusion matrix
}
