% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_binary_ratings}
\alias{as_binary_ratings}
\title{Convert to binary ratings}
\usage{
as_binary_ratings(raw_ratings, in_class)
}
\arguments{
\item{raw_ratings}{A data frame where the first column is subject_id,
the second column is a rating for that subject, and an optional third
column gives the rater_id}

\item{in_class}{A vector with the rating values that comprise the in-class. For
example, for a 1-5 scale, you might pick 1:3 to be the in-class, so that
the result is a binary distinction between {1,2,3} and {4,5}.}
}
\value{
A data frame with columns subject_id, rating, and rater_id if
that information is included.
}
\description{
The package functions require a specific format for the ratings
data, e.g. ratings must be binary and the column names must be subject_id,
rating, and (optionally) rater_id, followed by any other columns.
}
