% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{as_cat_ratings}
\alias{as_cat_ratings}
\title{Turn ratings into cat_ratings with default parameters}
\usage{
as_cat_ratings(ratings, labels = NULL, K = NULL)
}
\arguments{
\item{ratings}{A ratings data frame, with subject_id, rating, rater_id}

\item{labels}{An optional K-vector of character strings}

\item{K}{An optional vector length; the number of categories}
}
\value{
A cat_ratings object
}
\description{
Turn ratings into cat_ratings with default parameters
}
\details{
Providing K may be necessary if there are missing values in the
ratings, so that some classes 1..K don't appear at all.
}
