% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_count_index}
\alias{as_count_index}
\title{convert from long form to count index}
\usage{
as_count_index(ratings)
}
\arguments{
\item{ratings}{A dataframe with subject_id and binary rating}
}
\value{
A data frame with columns subject_id, N_r, and N_c
}
\description{
convert from long form to count index
}
