% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_rating_params}
\alias{as_rating_params}
\title{convert from ratings to rating_params}
\usage{
as_rating_params(ratings, params)
}
\arguments{
\item{ratings}{A dataframe with subject_id and binary rating}

\item{params}{A `params` object with t, a, p parameters}
}
\value{
A data frame with columns subject_id, rating, t, a, p
}
\description{
This is a convenience function for adding average
parameters to a ratings set, e.g. for use with the calibration functions.
Currently this only works with the basic t-a-p model.
}
