% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{as_rating_params_cat}
\alias{as_rating_params_cat}
\title{cat_ratings to rating_params}
\usage{
as_rating_params_cat(cat_ratings)
}
\arguments{
\item{cat_ratings}{A cat_ratings object after parameter estimation}
}
\value{
a data frame with ratings, ids, and parameters
}
\description{
cat_ratings to rating_params
}
\details{
This is a convenience function for joining the pieces of
information in cat_ratings into one data frame. The number of labels
must match the number of unique values of rating and will be matched
in ascending order of the ratings.
}
