% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{avg_params_cat}
\alias{avg_params_cat}
\title{Average params for a cat_ratings object}
\usage{
avg_params_cat(cat_ratings)
}
\arguments{
\item{cat_ratings}{A cat_ratings object}
}
\value{
a list with average t, a, p
}
\description{
Average params for a cat_ratings object
}
\details{
This is a convenience function for collapsing the complexity of a
cat_ratings$subjects and cat_ratings$raters parameters into averages over
the three parameters t, a, and p, where t and p are K-vectors.
}
