% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{bernoulli_trial}
\alias{bernoulli_trial}
\title{bernoulli trial}
\usage{
bernoulli_trial(p)
}
\arguments{
\item{p}{a vector of probabilities of success}
}
\value{
a vector of 0s and 1s
}
\description{
Simulate a Bernoulli trial with a given probability
}
