% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{check_convergence_cat}
\alias{check_convergence_cat}
\title{Compare two average parameter sets for abs difference}
\usage{
check_convergence_cat(p1, p2, tolerance)
}
\arguments{
\item{p1}{Parameter set (t, a, p) with t and p k-vectors}

\item{p2}{A second parameter set}

\item{tolerance}{What absolute difference to require?}
}
\description{
Compare two average parameter sets for abs difference
}
