% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{clamp_and_normalize_list_column}
\alias{clamp_and_normalize_list_column}
\title{Keep parameters within bounds}
\usage{
clamp_and_normalize_list_column(df, col, eps = 1e-12)
}
\arguments{
\item{df}{A data frame with a list-col}

\item{col}{The name of the column to be checked. It should be
K-vectors wrapped in lists.}
}
\value{
The list column is nudged up to 0 or down to 1 as needed and
then renormed to sum to 1
}
\description{
Keep parameters within bounds
}
