% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{compare_model_params}
\alias{compare_model_params}
\title{Compare model parameters}
\usage{
compare_model_params(
  rating_params_1,
  rating_params_2,
  model_names = c("model 1", "model 2"),
  vars = c("t", "a", "p", "bias"),
  markers = "points",
  scales = "fixed"
)
}
\arguments{
\item{rating_params_1}{output of either `fit_ratings()` or `fit_ratings_mcmc()`}

\item{rating_params_2}{output of either `fit_ratings()` or `fit_ratings_mcmc()`}

\item{model_names}{a character vector of length 2 with the names of the models}

\item{vars}{Some subset of "t", "a", "p", "bias" to include in the output plot,
defaults to all of them.}

\item{markers}{One of "points" or "labels", defaulting to points}

\item{scales}{Either "free" or "fixed", defaulting to fixed. Fixed scales
show the whole [0,1] range for all variables, while free scales adapt to the
data.}
}
\value{
A ggplot object with the comparison of the two models
}
\description{
Given two sets of model parameters as `rating_params` data frames,
plot the t,a,p coefficients for each to make a visual comparison.
}
