% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{compare_model_to_observed}
\alias{compare_model_to_observed}
\title{compare model to observed}
\usage{
compare_model_to_observed(model_params, observed_data = NULL, n_sims = 30)
}
\arguments{
\item{model_params}{The parameters of the model, a list with t, a0, a1, p0, p1,
OR a rating_params data frame with columns subject_id, rating, rater_id, t, a, p}

\item{observed_data}{frame of binary ratings or counts, which must either (1) have
columns SubjectID__and rating (0 or 1) in long format, or (2) have columns N_r,
N_c, and n (the count) to give the count distribution over subjects, respectively.
If rating_params is included as the model, then ratings can be omitted}

\item{n_sims}{How many sims to run to generate the modeled distribution? This
only gets used if rating_params are provided. For counts, exact calculations
are used.}
}
\value{
A dataframe with model fit statistics
}
\description{
compare model to observed
}
\details{
This function compares the model to observed data by aggregating the observed and
modeled distributions of class 1 counts by subject. It cannot accomodate individual-level
parameters like t_i, a_j, or p_j. Therefore, no RaterID__ is required.
}
