% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{compare_simulated_to_observed}
\alias{compare_simulated_to_observed}
\title{simulate data to compare to observed}
\usage{
compare_simulated_to_observed(
  n_iter = 100,
  observed_data,
  params,
  output = "plot"
)
}
\arguments{
\item{n_iter}{number of iterations to simulate, defaults to 100}

\item{observed_data}{data frame of observed data, which must have a fixed number of raters,
and columns RaterID__ (integer index) and rating (0 or 1)}

\item{params}{list of parameters for the simulation: n_subjects, n_raters, t, a0, a1, p0, p1}

\item{output}{If "data" a dataframe with the comparison data, or if "plot" a ggplot object (default)}
}
\value{
A dataframe with the observed and simulated frequencies of ratings
}
\description{
simulate data to compare to observed
}
