% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{count_ratings}
\alias{count_ratings}
\title{Count ratings}
\usage{
count_ratings(raw_ratings, in_class, summarize = FALSE)
}
\arguments{
\item{summarize}{If FALSE returns the count_index, which has one row per
subject, with N_r, N_c. If TRUE, returns counts, comprising the unique
pairs of N_r,N_c and how many times they occur n. Defaults to FALSE, since
some functions expect the count_index.}
}
\value{
A data frame with one row per subject, with the number of raters N_r,
and N_c the  number of raters who rated in the in-class range. If summarize
is TRUE, then the condensed form is returned.
}
\description{
The functions in this package require a special form for the
ratings data. This function wraps `as_binary_ratings` and either
`as_counts` or `as_count_index` to properly format the raw data.
There are two ways to summarize this data: one is to
provide row-wise counts of N_r (number of raters) and N_c (number of class 1
ratings), which optionally includes a RaterIR__. The other is to summarize
that information with a column n to indicate how many times that pair of N_r,
N_c occurs, which will drop any rater ID. The longer form is
referred to as a `count_index`, and the summarized form as `counts`. There
are support functions included to check validity and convert as needed.
}
