% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{dk_horizon}
\alias{dk_horizon}
\title{Dunning-Krueger Threshold}
\usage{
dk_horizon(N_s, N_r, a = 0, tp = 0.5, use_fleiss = TRUE, n_sim = 500)
}
\arguments{
\item{N_s}{number of subjects}

\item{N_r}{average number of raters per subject}

\item{a}{The "null value" of a to test against, usually zero (default)}

\item{tp}{The value of t and p, which must be equal. This is used for the
simulations as well as the exact coefficient model when use_fleiss = FALSE.}

\item{use_fleiss}{If TRUE, use the Fleiss kappa to estimate accuracy, otherwise
the exact formula is used. Default is TRUE. There are limitations to the
exact formula.}

\item{n_sim}{number of simulations to run, defaulting to 100}
}
\value{
A list with 50%, 90%, 95%, and 98% quantiles for a.
}
\description{
Estimates the Dunning-Kruger horizon as
a percentile of the estimate for accuracy when it is actually
zero.
}
\details{
The function uses the `simulate_exact_fit()` function to estimate the
}
