% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{est_signal_noise}
\alias{est_signal_noise}
\title{Signal and noise}
\usage{
est_signal_noise(params)
}
\arguments{
\item{params}{the three average t-a-p parameters in a named list. If vectors
are provided instead, they will be averaged.}
}
\value{
a data frame with columns for information from sources signal, noise,
and the total log likelihood per rating
}
\description{
Signal and noise
}
