% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{estimate_a_p_cat}
\alias{estimate_a_p_cat}
\title{Estimate a and p for K categories}
\usage{
estimate_a_p_cat(rating_params_cat, K)
}
\arguments{
\item{rating_params_cat}{ratings with vector t added to
subjects}

\item{K}{the number of categories}
}
\value{
A 1-row data frame(a, p) with updated params
}
\description{
Estimate a and p for K categories
}
