% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{estimate_aj_pj}
\alias{estimate_aj_pj}
\title{generate rater parameters}
\usage{
estimate_aj_pj(rating_params)
}
\arguments{
\item{rating_params}{A rating_params dataframe
rating, and rater_id (a `ratings` data frame). See the `as_binary_ratings()` function.}
}
\value{
a rating_params data frame with one row per rater giving t,a,p statistics for each
}
\description{
generate rater parameters
}
