% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{estimate_C}
\alias{estimate_C}
\title{Empirical confusion matrix for K-categories}
\usage{
estimate_C(rating_params_cat, K, normalize = TRUE)
}
\arguments{
\item{rating_params_cat}{ratings with vector t added to
subjects and a, vector p}

\item{K}{the number of categories}

\item{normalize}{if TRUE normalizes sum(C) = 1}
}
\value{
The estimated C matrix
}
\description{
Empirical confusion matrix for K-categories
}
\details{
This function is designed to be used inside data
grouped by rater if desired.
}
