% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_model_estimation.R
\name{estimate_params_mcmc}
\alias{estimate_params_mcmc}
\title{Get t-a-p coefficients from MCMC draws}
\usage{
estimate_params_mcmc(model_draws)
}
\arguments{
\item{model_draws}{Use `fit_counts_mcmc` or `fit_ratings_mcmc` to get the
fitted model, then use the `draws()` method to get the draws, e.g.
`fitted_model$draws()`. Don't forget the parentheses--this is a function
call.}
}
\value{
a data frame with var as the parameter name, with columns for
mean (avg), 5th percentile (p05), 25th percentile (p25), median, 75th
percentile (p75), 95th percentile (p95), mode1, mode2, and standard deviation.
}
\description{
Get t-a-p coefficients from MCMC draws
}
