% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{estimate_rater_parameter_error}
\alias{estimate_rater_parameter_error}
\title{recover rater parameters from simulated data}
\usage{
estimate_rater_parameter_error(rating_params, n_sim = 10)
}
\arguments{
\item{rating_params}{The ratings data with subject_id, rating, and rater_id
columns in addition to the parameter estimates. This normally comes from  a
solver like `fit_ratings(ratings)` or `fit_ratings_mcmc(ratings)$rating_params`}

\item{n_sim}{The number of simulations to run, defaulting to 10, since it can
be time-consuming.}
}
\value{
a list with plot_a = box plot of the accuracy estimates, plot_p for
the p_j parameters, and rater_params = a dataframe with simulation values for
each a_j, p_j parameter.
}
\description{
recover rater parameters from simulated data
}
