% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{estimate_ti}
\alias{estimate_ti}
\title{Truth probabilities by Class 1 rate for each subject}
\usage{
estimate_ti(rating_params)
}
\arguments{
\item{rating_params}{A data frame from `fit_ratings()` or `fit_ratings_mcmc()`}
}
\value{
a data frame with subject_id, t_i, C0, and C1 columns. The latter two
are the probability estimates of the classes for each subject, where
t_i = Pr(Class 1) = C1/(C0 + C1).
}
\description{
Given a rating_params data frame, which contains the t, a, and p
parameters for each rating, as well as the rating, recalculate the t_i
for each subject using the general (multinomial) calculation.
}
\details{
In the model, some ratings of Class 1 are generated from true Class 1 cases
and the rest from true Class 0 cases. The derivation mimics that found
in the MCMC model t_i-a_j-p_j.stan. The function `estimate_tu()` does the
same thing for the counts of ratings, where we assume fixed values for
each of the parameters (fully pooled model).
}
