% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{estimate_ti_cat}
\alias{estimate_ti_cat}
\title{Truth probabilities by Class 1 rate for each subject}
\usage{
estimate_ti_cat(cat_ratings)
}
\arguments{
\item{cat_ratings, }{with the appropriate structure. See `generate_sample_ratings_cat()`}
}
\value{
a cat_ratings object with updated t vectors
}
\description{
Given a rating_params data frame, which contains the t, a, and p
parameters for each rating, as well as the rating, recalculate the t_i
for each subject using the general (multinomial) calculation.
}
