% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{estimate_tu}
\alias{estimate_tu}
\title{Truth probabilities by Class 1 rate}
\usage{
estimate_tu(counts, params)
}
\arguments{
\item{counts}{A data frame with columns N_r, and N_c. This might come from
`count_ratings()` with or without summarize = TRUE.}

\item{params}{A named list with t, a, and p estimates or a data frame with
a single row and columns named t, a, and, p. The function can also acommodate
parameter lists with a0,a1 and/or p0,p1 instead of a and p, respectively.}
}
\value{
a data frame with the original columns plus columns C0 and C1
for the asymptotic rates of ratings of those classes, and a column  t_u for
the estimate of the true rate of Class 1 given counts.
}
\description{
Given a data frame of (N_r, N_c) pairs and t-a-p parameters
add a column to the dataframe with the probability of Class 1 for each
}
\details{
In the model, some ratings of Class 1 are generated from true Class 1 cases
and the rest from true Class 0 cases. Given the parametes we can find the
distributions of these two rating patterns, from which we can estimate the
fraction of each rating sum (e.g. 3 of 5 raters assigning class 1) that is
due to true Class 1 cases. We use that as an estimate for the probability of
the true class being 1.
}
