% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{exact_accuracy_coefs}
\alias{exact_accuracy_coefs}
\title{Estimate the coefficients for the closed-form accuracy model}
\usage{
exact_accuracy_coefs(N_r, tp)
}
\arguments{
\item{N_r}{the number of raters (a constant)}

\item{tp}{the t and p parameters, which are assumed equal (unbiased raters)}
}
\value{
a data frame with the coefficients for the closed-form model of
accuracy as a function of the count proportions observed in data
}
\description{
Given the number of raters N_r, the true rate of Class 1 t, and the random
assignment rate p, estimate the coefficients for the closed-form model of
accuracy as a function of the count proportions observed in data
}
\details{
The functions that start with "exact_" are used for closed-form
calculation of accuracy. You'll get a warning if there are more than
thirty raters, because colinearity makes the regression unstable above that
point.
}
