% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{exact_combine_counts}
\alias{exact_combine_counts}
\title{Apply symmetry to halve the length of count probabilities (private)}
\usage{
exact_combine_counts(count_prob)
}
\arguments{
\item{count_prob}{a vector of count probabilities from `exact_count_probabilites()`}
}
\value{
a vector of count probabilities with the second half reflected and
summed with the first half, accounting for even/odd cases.
}
\description{
Apply symmetry to halve the length of count probabilities (private)
}
\details{
This function is not exported, since it's only used with
`exact_count_probabilites()` and `exact_count_range()`. The functions that
start with "exact_" are used for closed-form calculation of accuracy.
}
