% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{exact_count_probabilites}
\alias{exact_count_probabilites}
\title{get count probabilities (private)}
\usage{
exact_count_probabilites(N_r, t, a, p)
}
\arguments{
\item{N_r}{the number of raters}

\item{t}{the true rate of Class 1}

\item{a}{rater accuracy}

\item{p}{random assigment rate for inaccurate ratings}
}
\value{
a vector of probabilities for each count n0, n1, ... that has
been reflected and summed to account for the symmetry of the distribution
}
\description{
Given the number of raters and t,a,p parameters, generate the probability
distribution for the number of counts of Class 1 ratings per subject.
}
\details{
The functions that start with "exact_" are used for closed-form
calculation of accuracy.
}
