% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{exact_count_range}
\alias{exact_count_range}
\title{Generate inputs for the regression to find coefficients for closed-form (private)}
\usage{
exact_count_range(N_r, t, p)
}
\arguments{
\item{N_r}{the number of raters}

\item{t}{the true rate of Class 1}

\item{p}{random assigment rate for inaccurate ratings}
}
\value{
a list with two elements: a vector of accuracy values, which is the
dependent variable for the regression, and the matrix of predictors comprising
the exact count probabilities for each accuracy value.
}
\description{
Generates a sequence of accuracy values from 0 to 1 in increments of .001 and
associated exact distributions of the count probabilities from the t-a-p
model with the provided number of raters N_r, t, and p.
}
\details{
The functions that start with "exact_" are used for closed-form
calculation of accuracy.
}
