% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{exact_fill_missing_vals}
\alias{exact_fill_missing_vals}
\title{Fill in zeros for missing ratings counts (private)}
\usage{
exact_fill_missing_vals(counts)
}
\arguments{
\item{counts}{A data frame resulting from `as_counts(ratings)`}
}
\value{
A data frame with all possible counts from 0 to N_r, with zeros added
where necessary.
}
\description{
To use the coefficients generated from `exact_accuracy_coefs()` we need a
complete set of count proportions. For real data sets, especially with
larger numbers of raters, we might have counts that don't appear in the results.
This function fills in those blanks by inserting zeros.
}
