% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_params}
\alias{expand_params}
\title{Expand t-a-p params to t-a0a1-p0p1}
\usage{
expand_params(params)
}
\arguments{
\item{params}{A list with the t-a-p parameters}
}
\value{
An expanded params list
}
\description{
Expand t-a-p params to t-a0a1-p0p1
}
