% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{expected_krits_per_rating_cat}
\alias{expected_krits_per_rating_cat}
\title{expected_krits_per_rating_cat: Average entropy per rating (base 1/K) from the model.}
\usage{
expected_krits_per_rating_cat(params, uniform_t = FALSE)
}
\arguments{
\item{params}{A list with t, a, and p}

\item{uniform_t}{If TRUE, t is set to a uniform distribution to eliminate
the effect of sampling from true classes}
}
\value{
A numeric scalar: average entropy per rating in the range [0, 1].
}
\description{
expected_krits_per_rating_cat: Average entropy per rating (base 1/K) from the model.
}
\seealso{
\code{\link{bits_per_rating}}
}
