% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_model_estimation.R
\name{extract_param_draws_mcmc}
\alias{extract_param_draws_mcmc}
\title{Extract the parametersfrom a stan draws object}
\usage{
extract_param_draws_mcmc(model_draws)
}
\arguments{
\item{model_draws}{a stan draws object}
}
\value{
a data frame with one row per draw and one column per parameter
}
\description{
Create a data frame from a draws object, with one row per draw
and one column per parameter
}
\details{
This is a useful utility for combining the chains to assemble a
data frame with parameters as column names and all the draws underneath.
Was called extract_vars() in a prior version of the package.
}
