% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract the draws from a stan draws object}
\usage{
extract_vars(model_draws)
}
\arguments{
\item{model_draws}{a stan draws object}
}
\value{
a data frame with one row per draw and one column per parameter
}
\description{
Create a data frame from a draws object, with one row per draw
and one column per parameter
}
