% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{find_solution}
\alias{find_solution}
\title{Fit the t-a-p model to binary ratings}
\usage{
find_solution(
  counts,
  init_params,
  method = "EM",
  max_steps = 100,
  tolerance = 1e-04
)
}
\arguments{
\item{counts}{a summary data frame with columns N_r, N_c, representing replications
of unique pairs of (N_r, N_c). This is the output of `count_ratings` with
`summarize = TRUE`
raters and counts of raters who rated class 1.}

\item{init_params}{initial values for t, a, and p}

\item{method}{optimization method, defaults to "EM" for expectation maximization,
or "optim" for a more general optimization method using R's method}

\item{max_steps}{maximum number of iterations for the EM optimization, defaults to 100}

\item{tolerance}{convergence tolerance for the EM optimization, defaults to 1e-4}
}
\value{
a data frame with columns for parameters: t, a, p, and log likelihood of the solution
}
\description{
Fit the t-a-p model to binary ratings
}
