% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{find_solution_em}
\alias{find_solution_em}
\title{Fit the t-a-p model to counts using EM algorithm}
\usage{
find_solution_em(counts, max_steps = 100, tolerance = 1e-04)
}
\arguments{
\item{counts}{a data frame with columns N_r, N_c, and n}

\item{max_steps}{the maximum number of iterations to run, defaults to 100}

\item{tolerance}{the convergence tolerance, defaults to 1e-4}
}
\value{
a data frame with columns t, a, p, ll, and degenerate
}
\description{
Fit the t-a-p model to counts using EM algorithm
}
