% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{fit_counts}
\alias{fit_counts}
\title{iterative attempt to fit model from counts}
\usage{
fit_counts(counts, method = "EM")
}
\arguments{
\item{counts}{a data frame with columns N_r, N_c, and n,  representing number
of raters and counts of raters who rated class 1 with multiplicities n. This
can be created from `as_counts(ratings)`}

\item{method}{the method to use for optimization, defaults to "EM" for
expectation-maximization, or use "optim" for a direct optimization}
}
\value{
a data frame with columns for parameters: t, a, p, and log likelihood
of the solution, and a flag for degeneracy
}
\description{
Fit the t-a-p model to counts. Start with the the middle of the
parameter space in looking #' for a solution. If it's degenerate it looks
around the edges of the space for a non-degenerate solution.
}
