% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_model_estimation.R
\name{fit_counts_mcmc}
\alias{fit_counts_mcmc}
\title{Generate t-a-p statistics from counts}
\usage{
fit_counts_mcmc(
  counts,
  stan_model = "t-a-p",
  quiet = FALSE,
  mcmc_chains = 3,
  mcmc_iter = 1000,
  mcmc_warmup = 300,
  mcmc_cores = 3
)
}
\arguments{
\item{counts}{a data frame with *summarized* ratings counts, one row per
pair (N_r, N_c) with a column n for the count of occurrences. It
must include columns N_r, N_c, and n. This is the output of `as_counts(ratings)`.}

\item{stan_model}{a stan model suitable for count data. One of "t-a-p" (default)
or "t-a0a1-p" for separate accuracies for the two classes, or "t-a0a1-p0p1"
to also include separate bias parameters. You can retrieve these models, e.g.
with  `system.file("inst/stan", "t-a-p.stan", package = "tapModel")`}

\item{quiet}{If TRUE, discards the normal status updates. This is useful
if generating a qmd file, for example. Defaults to FALSE.}

\item{mcmc_chains}{number of MCMC chains to run, defaulting to 3}

\item{mcmc_iter}{number of MCMC iterations per chain, defaulting to 1000}

\item{mcmc_warmup}{number of MCMC burn-in iterations, defaulting to 300}

\item{mcmc_cores}{number of cores to use for MCMC, defaulting to 3}
}
\value{
a list with params = estimated parameters, and fitted_model = the
fitted model. You can use `shinystan::launch_shinystan(fitted_model)` to
explore the model fit.
}
\description{
Generate t-a-p statistics from counts
}
\details{
Requires `library(cmdstanr)` and `library(LaplacesDeamon)`. If you're
looking for models that assign individual truth and/or accuracy parameters,
first estimate these with `rating_params = fit_ratings(ratings)`, then use
 `fit_ratings_mcmc(rating_params)` instead of this function.
}
