% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{fit_ratings}
\alias{fit_ratings}
\title{fit ratings to generate t_i, a_j, p_j for each rating}
\usage{
fit_ratings(
  ratings,
  params = NULL,
  max_iterations = 10,
  ll_threshold = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{ratings}{a data frame with columns subject_id, rating, and rater_id}

\item{params}{optional average t-a-p parameters to use for the fit, or (if NULL)
automatically generated}

\item{max_iterations}{The maximum number of iterations to run, defaults to 10}

\item{ll_threshold}{The threshold for log likelihood convergence, defaults to .01}

\item{verbose}{report out iterations?}
}
\value{
A data frame with the original ratings data augmented with
parameters for each subject, rater, and rating
}
\description{
fit ratings to generate t_i, a_j, p_j for each rating
}
