% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{fit_ratings_cat}
\alias{fit_ratings_cat}
\title{Fit categorical t-a-p model (full hierarchical EM)}
\usage{
fit_ratings_cat(cat_ratings, max_iter = 20, tol = 1e-04)
}
\arguments{
\item{cat_ratings}{A cat_ratings object (see generate_sample_ratings_cat()).}

\item{max_iter}{Maximum number of EM iterations (default 50).}

\item{tol}{Convergence tolerance on log-likelihood (default 1e-6).}
}
\value{
Updated cat_ratings object with fitted subject and rater parameters.
}
\description{
Fit categorical t-a-p model (full hierarchical EM)
}
