% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_model_estimation.R
\name{fit_ratings_mcmc}
\alias{fit_ratings_mcmc}
\title{fit ratings with Bayesian MCMC}
\usage{
fit_ratings_mcmc(
  ratings,
  stan_model = "t_i-a_j-p_j",
  quiet = FALSE,
  mcmc_chains = 3,
  mcmc_iter = 1000,
  mcmc_warmup = 300,
  mcmc_cores = 3,
  mcmc_prior_sd = c(1, 1, 1)
)
}
\arguments{
\item{ratings}{a ratings data frame OR a one augmented with parameter estimates,
the output of `fit_ratings()`. If the former, initial estimates will be obtained
by first running `fit_counts_mcmc` on the counts of ratings, and then
`fit_ratings()` used to get starting estimates for the
individual parameters.}

\item{quiet}{If TRUE, discards the normal status updates. This is useful
if generating a qmd file, for example. Defaults to FALSE.}

\item{mcmc_chains}{number of MCMC chains to run, defaulting to 3}

\item{mcmc_iter}{number of MCMC iterations per chain, defaulting to 1000}

\item{mcmc_warmup}{number of MCMC burn-in iterations, defaulting to 300}

\item{mcmc_cores}{number of cores to use for MCMC, defaulting to 3}

\item{mcmc_prior_sd}{use the priors from the t-a-p estimates as means for the
parameters and the standard deviations specified here as a 3-vector in order
t, a, p, defaulting to c(1,1,1). Use c(.01,.05,.05) for a very tight constraint.
If any of these is set to 1 or more, the prior for that parameter is set to
a uniform distribution on [0,1].}

\item{model}{the model to fit, one of "t_i-a_j-p_j" (default) or "t-a_j-p_j"}
}
\value{
a list with rating_params =data frame with the original ratings
data augmented with parameter estimates using the mode1 of the posterior
distribution, and fitted_model = the fitted model. Note that if you
want more detailed summary information about the parameters, you can use
`tapModel::estimate_params_mcmc(fitted_model$draws())` where the
fitted_model is the output of this function.
}
\description{
fit ratings with Bayesian MCMC
}
\details{
This function takes the estimates from the three parameter t-a-p
as initial values, and applies priors that are specified by the user with
the mcmc_prior_sd parameter.
}
