% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{fit_ratings_mcmc_deprecated}
\alias{fit_ratings_mcmc_deprecated}
\title{Generate t-a-p statistics from ratings -- NOT USED CURRENTLY}
\usage{
fit_ratings_mcmc_deprecated(
  ratings,
  compiled_model,
  output = "draws",
  lambda = 30
)
}
\arguments{
\item{ratings}{a data frame that must include the columns SubjectID, RaterID,
and rating that are necessary for the model. For example, if running the
t_i-a-p model, the data frame must include the columns SubjectID and rating.}

\item{compiled_model}{a compiled stan model suitable for this data}

\item{output}{Specify the output as "model" for the complete stan output object,
or "draws" for the draws from the model (the default), or "params" to get the
paramter draws in a data frame.}

\item{lambda}{A non-negative integer to specify the steepness of the logistic
function that creates a soft step. Defaults to 30. Not used in all models.}
}
\value{
Depending on the output parameter, either the whole stan model output,
a posterior stan \code{draws} object, or a data frame with the parameter draws.
}
\description{
This option is for models that assign individual truth and/or accuracy
parameters, so called "random effects" models or "hierarchical" models.
The extra detail requires that the individual ratings, rather than counts
per subject, must be provided. #'
}
\details{
If you have counts instead of raw ratings, use `fit_counts_mcmc`. This
function will work with stan models that accumulate log likelihood from
each rating. This requires more indexing, and hence the
requirement for the Rater and/or Subject IDs. See `t_i-a-p.stan` for an example.
was fit_random_tap_model
}
