% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_model_estimation.R
\name{fit_ratings_mcmc_latent}
\alias{fit_ratings_mcmc_latent}
\title{fit ratings with Bayesian MCMC using latent parameters}
\usage{
fit_ratings_mcmc_latent(
  rating_params,
  quiet = FALSE,
  mcmc_chains = 3,
  mcmc_iter = 1000,
  mcmc_warmup = 300,
  mcmc_cores = 3,
  mcmc_prior_sd = c(0.1, 0.05, 0.05)
)
}
\arguments{
\item{rating_params}{a ratings data frame augmented with parameter estimates,
the output of `fit_ratings()`}

\item{quiet}{If TRUE, discards the normal status updates. This is useful
if generating a qmd file, for example. Defaults to FALSE.}

\item{mcmc_chains}{number of MCMC chains to run, defaulting to 3}

\item{mcmc_iter}{number of MCMC iterations per chain, defaulting to 1000}

\item{mcmc_warmup}{number of MCMC burn-in iterations, defaulting to 300}

\item{mcmc_cores}{number of cores to use for MCMC, defaulting to 3}
}
\value{
a list with rating_params =data frame with the original ratings
data augmented with parameter estimates, and fitted_model = the fitted model.
}
\description{
fit ratings with Bayesian MCMC using latent parameters
}
\details{
This function takes the estimates from the three parameter t-a-p
and uses them as initial values to estimate all the paramters on latent
scales with regularization. Instead of parameters in [0,1], they are
on the real line, and mapped to [0,1] with the logistic function. Each of
t, a, and p have an average parameter with offsets for the individual
parameters of the form tau_i = tau_avg + sd_tau*tau_i, where tau_avg, tau_i,
and sd_tau have normal(0,1) priors. The idea is that the model can optimize
between average and freely roaming individual parameters via shrinkage,
to optimize likelihood without overfitting the data.
}
