% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{fit_tap_model}
\alias{fit_tap_model}
\title{Generate t-a-p statistics from counts}
\usage{
fit_tap_model(counts, compiled_model)
}
\arguments{
\item{compiled_model}{a compiled stan model suitable for this data}

\item{ratings}{a data frame that must include columns N_r and N_c, the number
of raters and ratings of class 1 for each subject}
}
\value{
a posterior stan \code{draws} object
}
\description{
Generate t-a-p statistics from counts
}
\details{
The model must be able to accommodate varying numbers of raters per
subject
}
