% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{fix_domain}
\alias{fix_domain}
\title{Ensure proper [0,1] domain for a vector}
\usage{
fix_domain(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
numerical vector with values in [0,1], rounded where necessary
}
\description{
Ensure proper [0,1] domain for a vector
}
