% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{fleiss_kappa}
\alias{fleiss_kappa}
\title{Fleiss kappa}
\usage{
fleiss_kappa(counts)
}
\arguments{
\item{counts}{a data frame with columns N_r and N_c, representing number of
raters for each subject and number of raters who rated class 1, and n for the
number of these (N_r, N_c) pairs. Optionally the row counts can be used, and the
function will convert}
}
\value{
a data frame with one row and columns for parameters: t, a, p, log likelihood: ll,
and a true/false flag for degenerate solutions.
}
\description{
Function to calculate Fleiss' kappa for binary data
}
