% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{fleiss_kappa_cat}
\alias{fleiss_kappa_cat}
\title{Fleiss' kappa from cat_ratings}
\usage{
fleiss_kappa_cat(cat_ratings)
}
\arguments{
\item{cat_ratings}{A list object returned by generate_sample_ratings_cat()
with elements ratings, subjects, raters, K, and labels}
}
\value{
cat_ratings with updated t, a, p, ll, and degenerate flag
}
\description{
Compute Fleiss' kappa for categorical ratings (K >= 2)
}
