% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{format_data}
\alias{format_data}
\title{format rater data}
\usage{
format_data(
  df,
  subject_id = NULL,
  category = NULL,
  ratings,
  data_type = "categories"
)
}
\arguments{
\item{df}{A data frame with one or more columns for raters and one or
more columns for ratings.}

\item{subject_id}{The optional name for the column with the subject ID, or
NULL the default, if there is none. The long format setting for
data_type requires a subject ID.}

\item{category}{The optional name for the column with the category of the
rating, or NULL, the default, if all the ratings are of a single thing.}

\item{ratings}{The names of the columns with the ratings. If the data is in}

\item{data_type}{What do the columns comprise? There are three data types
supported.
Use "categories" when each row is a subject ID and one or more categories,
with an optional rater ID. For example, a rating set on a service might have
columns for friendliness, cleanliness, and speed, with ratings in each of
those columns. Use "raters" when each row is a subject ID and two or more
raters of that subject, with an optional category column. Here, the ratings
for each rater appears under the column for that rater. The "narrow" format
is when each row is a unique combination of subject and rater, with an
optional category column. This requires the subject ID to exist.
'}
}
\value{
A data frame with a column for subject ID and one or more columns for
}
\description{
Standardize rater data to have a column with subject ID and one or more
columns for aspects being rated for that subject.
}
\details{
For the subject_id and category, the string "[None]" can be substituted
for NULL. This is included for compatability with the app.
}
