% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{format_raw_ratings}
\alias{format_raw_ratings}
\title{Format raw rating data for analysis}
\usage{
format_raw_ratings(
  raw_ratings,
  rating_cols,
  rating_colnames_type,
  subject_id_col = NULL,
  rater_id_col = NULL,
  prune = FALSE
)
}
\arguments{
\item{raw_ratings}{A data frame with one or more columns for raters and one or
more columns for ratings.}

\item{subject_id_col}{The optional name for the column with the subject ID, or
NULL or "[None]" if there is not one. Defaults to NULL.}

\item{rater_id_col}{The optional name for the column with the rater ID, or
NULL or "[None]" if there is not one. Defaults to NULL.}

\item{prune}{If TRUE, remove any columns except for subject_id, rating, and
rater_id, if present. Defaults to FALSE, which will allow other descriptive columns
to be included, e.g. categories of ratings. You shouldn't prune the ratings
if you have more than one category, because normally you'll want to analyze
one category at a time.}

\item{ratings_cols}{The required names of the column(s) that contain ratings.}

\item{rating_colname_type}{The header names of the ratings column(s) must be one of
"rating", rater", "subject", or "category." The first, "rating" is for simple
cases with a single column for rating data. The others are for "wide" data
cases where the ratings are spread over two or more columns For example,
if you have columns named rater1, rater2, and rater3, each with ratings in them,
the value of this parameter would be "rater."}
}
\value{
A list with the formatted ratings and a way to map these abstracted
values back to the original data. In the list, the element `ratings` is a
data frame in long format with one rating per row, identified by
subject (required), rater (optional) and other descriptive columns that may
be present, like category of ratings. The function
`as_binary_ratings(ordinal_ratings, in_class = ...)` can be used
to convert from nominal or ordinal to binary.
The data frame has columns with standardized names: in order
`subject_id`, `rating`, and `rater_id` (if present), followed by any other
columns that may be present.
}
\description{
Standardize rater data to ratings and subject ID, with optional columns.
}
\details{
This function assumes that the data has ratings in columsn not rows.
If your ratings are in rows, then you'll need to transpose and rename the
columns appropriately. There must be a subject identifier in the raw_data.
This can come from the names of the ratings columns, e.g. "subject1",
"subject2" as columns with ratings below, or the subject identifier can be a
separate column, e.g. "subject". If neither of these are present, the
formatting function will create a subject_id column from the row numbers.
If the ratings are in categories, then there must be a subject ID column
since there will presumably be a rater_id column (otherwise you can only have
 one rating per subject).
}
