% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ratings.R
\name{generate_exact_ratings}
\alias{generate_exact_ratings}
\title{Generate ratings that have a given distribution}
\usage{
generate_exact_ratings(
  N_s = 100,
  N_r = 5,
  params = list(t = 0.5, a = 0.7, p = 0.5)
)
}
\arguments{
\item{params}{t, a, and p, possibly including a0,a1,p0,p1 in a named list}

\item{n_subjects}{number of subjects}

\item{n_raters}{number of raters per subject on average}
}
\value{
a data frame with ratings and parameters, including the input t-a-p averages, the
random effect parameters p_i and a_j and the discrete random variables T_i, P_ij, and A_ij.
}
\description{
Generate ratings that have a given distribution
}
\details{
This function attempts to exactly replicate the density function
specified by the parameters. This can be useful in tests where you want to
avoid sampling effects.
}
