% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ratings.R
\name{generate_ti_aj_pj_ratings}
\alias{generate_ti_aj_pj_ratings}
\title{Random ratings with individual parameters}
\usage{
generate_ti_aj_pj_ratings(
  rating_params = NULL,
  param_list = NULL,
  use_avg_t = TRUE
)
}
\arguments{
\item{rating_params}{The output of `fit_ratings()` or NULL (defult)}

\item{param_list}{A list with probability vectors t, a, and p, or NULL.
Vectors a and p must be the same length. Defaults to NULL.}

\item{use_avg_t}{If TRUE, averages the t_i values for the ratings and uses that
instead of individual subject t_i parameters to generate ratings. This is a
more robust kind of simulation.}
}
\value{
A dataframe with columns subject_id, rating, and rater_id in addition
to the simulated T_i, A_ij, and P_ij values.
}
\description{
Random ratings with individual parameters
}
\details{
Exactly one of the two arguments should be provided. This is intended
to make it easy to simulate a data set with either the same parameters as a
model under study  or to specify a set of parameters from scratch. If the
param_list is provided, each rater will rate each subject.
}
