% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{generate_ti_aj_pj_ratings_cat}
\alias{generate_ti_aj_pj_ratings_cat}
\title{Generate ratings from a cat_ratings object}
\usage{
generate_ti_aj_pj_ratings_cat(cat_ratings, use_avg_t = TRUE)
}
\arguments{
\item{cat_ratings}{a cat_ratings object with subjects, raters, and ratings data frames}

\item{use_avg_t}{If TRUE, sets each subject's t vector to the average (default)}
}
\value{
A cat_ratings object with ratings generated from the t, a, p parameters
}
\description{
Generate ratings from a cat_ratings object
}
\details{
If use_avg_t is TRUE, then the t vector for each subject is set to the
average t vector over all subjects.
}
