% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{get_density}
\alias{get_density}
\title{faceted density plots for mcmc draws}
\usage{
get_density(x)
}
\arguments{
\item{x}{a numerical vector of parameter draws}
}
\value{
a data frame with x, y, and region columns
}
\description{
faceted density plots for mcmc draws
}
