% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{get_ordinal_tap}
\alias{get_ordinal_tap}
\title{tap optim over ordinal scale}
\usage{
get_ordinal_tap(ratings)
}
\arguments{
\item{ratings}{a data frame in long format with columns SubjectID__ and rating
where the ratings are for a single dimension, with multiple ordinal values,
for example water quality on a scale of 1 = poor, 2 = fair, 3 = good,
4 = excellent. If the ratings are character strings, the will be sorted
alphabetically. If the ratings are numeric, they will be sorted numerically.}
}
\value{
a data frame with columns for each cut point between ratings,
parameters: t, a, p, for the t-a-p model and the Fleiss kappa.
}
\description{
tap optim over ordinal scale
}
