% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{get_tap_stats}
\alias{get_tap_stats}
\title{Get t-a-p coefficients}
\usage{
get_tap_stats(model_draws)
}
\arguments{
\item{compiled_model}{The stan model for t-a0,a1-p_i}

\item{model_data}{A dataframe with columns: ID n_ratings, and sum_ratings.
ID is the subject identifier, n_ratings is the number of binary ratings,
and sum_ratings is the sum of the binary ratings.}
}
\value{
a list with the t-a-p coefficients. If there are two modes for a
these are returned as a0 and a1
}
\description{
Get t-a-p coefficients
}
