% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{information_bits}
\alias{information_bits}
\title{Information bits from a binary vector}
\usage{
information_bits(x)
}
\arguments{
\item{x}{a probability}
}
\value{
bits of information using Shannon entropy
}
\description{
Information bits from a binary vector
}
