% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_degenerate}
\alias{is_degenerate}
\title{Is the t-a-p model degenerate?}
\usage{
is_degenerate(params, threshold = 1e-06)
}
\arguments{
\item{params}{vector of t-a-p parameters}

\item{threshold}{threshold for degeneracy, defaults to x < 1e-6}
}
\value{
TRUE if the model is degenerate
}
\description{
given t-a-p parameters, return TRUE if the model is degenerate
}
