% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{iterative_optim}
\alias{iterative_optim}
\title{Iterative optimization for the t-a-p model}
\usage{
iterative_optim(counts, iterations = 1)

iterative_optim(counts, iterations = 1)
}
\arguments{
\item{counts}{a data frame with columns N_r and N_c, representing number of
raters and counts of raters who rated class 1.}

\item{iterations}{number of iterations to run. Currently this doesn't do
anything. It's there for future development.}
}
\value{
a data frame with columns for parameters: t, a, p, and log likelihood of the solution

a data frame with columns for parameters: t, a, p, and log likelihood
of the solution, and a flag for degeneracy
}
\description{
Start with the middle case and iterate to find the best solution

Start with the the middle of the parameter space in looking
for a solution. If it's degenerate it looks around the edges of the space
for a non-degenerate solution.
}
