% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_add}
\alias{lc_add}
\title{lc_add: Elementwise add of two list-columns}
\usage{
lc_add(lc1, lc2)
}
\arguments{
\item{lc1}{A list of numeric vectors.}

\item{lc2}{A list of numeric vectors (same length, same K).}
}
\value{
A list-column with elementwise sums.
}
\description{
lc_add: Elementwise add of two list-columns
}
