% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_bind}
\alias{lc_bind}
\title{lc_bind: Bind list-column into a matrix}
\usage{
lc_bind(lc)
}
\arguments{
\item{lc}{A list of numeric vectors (all same length).}
}
\value{
A numeric matrix with rows bound from the list-column.
}
\description{
lc_bind: Bind list-column into a matrix
}
