% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_dot}
\alias{lc_dot}
\title{lc_dot: Rowwise dot product of two list-columns}
\usage{
lc_dot(lc1, lc2)
}
\arguments{
\item{lc1}{First list of numeric vectors.}

\item{lc2}{Second list of numeric vectors.}
}
\value{
A numeric vector of dot products.
}
\description{
lc_dot: Rowwise dot product of two list-columns
}
