% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_fn}
\alias{lc_fn}
\title{lc_fn: Apply a function elementwise to list-column vectors}
\usage{
lc_fn(lc, fn)
}
\arguments{
\item{lc}{A list of numeric vectors.}

\item{fn}{A function applied to each vector.}
}
\value{
A list-column with fn applied to each vector.
}
\description{
lc_fn: Apply a function elementwise to list-column vectors
}
