% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_indicator}
\alias{lc_indicator}
\title{lc_indicator: Create one-hot indicator vectors}
\usage{
lc_indicator(idx, K)
}
\arguments{
\item{idx}{An integer vector with values 1..K.}

\item{K}{Number of categories.}
}
\value{
A list of one-hot vectors of length K.
}
\description{
lc_indicator: Create one-hot indicator vectors
}
