% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_mpy}
\alias{lc_mpy}
\title{lc_mpy: Multiply list-column by scalar or list-column}
\usage{
lc_mpy(lc1, x)
}
\arguments{
\item{lc1}{A list of numeric vectors.}

\item{x}{Either a numeric vector (scalars) or a list of numeric vectors.}
}
\value{
A list-column with elementwise products.
}
\description{
lc_mpy: Multiply list-column by scalar or list-column
}
