% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_norm}
\alias{lc_norm}
\title{lc_norm: Norm of list-column vectors}
\usage{
lc_norm(lc, type = c("l2", "l1"))
}
\arguments{
\item{lc}{A list of numeric vectors.}

\item{type}{Either "l1" (sum of abs) or "l2" (Euclidean).}
}
\value{
A numeric vector of norms.
}
\description{
lc_norm: Norm of list-column vectors
}
