% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_one}
\alias{lc_one}
\title{lc_one: Create one-filled list-column}
\usage{
lc_one(N, K)
}
\arguments{
\item{N}{Number of rows.}

\item{K}{Length of each vector.}
}
\value{
A list of N numeric vectors of length K, all one.
}
\description{
lc_one: Create one-filled list-column
}
