% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_subtract}
\alias{lc_subtract}
\title{lc_subtract: Elementwise subtraction of two list-columns}
\usage{
lc_subtract(lc1, lc2)
}
\arguments{
\item{lc1}{A list of numeric vectors.}

\item{lc2}{A list of numeric vectors (same length, same K).}
}
\value{
A list-column with elementwise lc1 - lc2.
}
\description{
lc_subtract: Elementwise subtraction of two list-columns
}
