% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_vsum}
\alias{lc_vsum}
\title{lc_vsum: Vertical sum of list-column vectors}
\usage{
lc_vsum(lc)
}
\arguments{
\item{lc}{A list of numeric vectors (all length K).}
}
\value{
A numeric vector of length K (the summed vector).
}
\description{
Collapse N K-vectors into a single K-vector by summing elementwise.
}
