% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{lc_zero}
\alias{lc_zero}
\title{lc_zero: Create zero-filled list-column}
\usage{
lc_zero(N, K)
}
\arguments{
\item{N}{Number of rows.}

\item{K}{Length of each vector.}
}
\value{
A list of N numeric vectors of length K, all zero.
}
\description{
lc_zero: Create zero-filled list-column
}
