% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{ll_per_subject}
\alias{ll_per_subject}
\title{log likelihood for each subject}
\usage{
ll_per_subject(rating_params)
}
\arguments{
\item{rating_params}{A data frame with binary subject_id, rating, and a
t, a, and p parameter for the rating.}
}
\value{
The rating_params data frame with column ll = log likelihood added.
The log base is e, so if you want bits divide by -log(2)
}
\description{
Given a rating_param data frame from `fit_ratings` or `fit_ratings_mcmc`,
return the log likelihood for each subject using a t_i-a_j-p_j model.
}
