% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LSE_R}
\alias{LSE_R}
\title{Log-sum-exponential}
\usage{
LSE_R(vec)
}
\arguments{
\item{vec}{a vector of numerics}
}
\value{
a scalar log(exp(vec[1]) + exp(vec[2]) ...)
}
\description{
Log-sum-exponential
}
\details{
See https://rpubs.com/FJRubio/LSE, where I got the code from.
Note that this isn't vectorized, so you need to use rowwise()
}
