% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{model_fit_comparison}
\alias{model_fit_comparison}
\title{fit a hierarchical t-a-p model to ratings data}
\usage{
model_fit_comparison(
  rating_params_1,
  rating_params_2 = NULL,
  model_names = c("model 1", "model 2"),
  n_sim = 200,
  vars = c("t", "a", "p", "c", "ll")
)
}
\arguments{
\item{rating_params_1}{output of either `fit_ratings()` or `fit_ratings_mcmc()`}

\item{rating_params_2}{output of either `fit_ratings()` or `fit_ratings_mcmc()`,
or NULL to use the average t-a-p model}

\item{model_names}{a character vector of length 2 with the names of the models}

\item{n_sim}{number of simulations to run, defaulting to 1000}

\item{vars}{Some subset of "t", "a", "p", "ll" to include in the output plot}
}
\value{
A list with the simulation data (sim_data) and a plot (plot)
comparing model fit using log likelihood in bits
}
\description{
fit a hierarchical t-a-p model to ratings data
}
