% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{picket_fence}
\alias{picket_fence}
\title{Find likelihood along a grid}
\usage{
picket_fence(counts)
}
\arguments{
\item{counts}{data frame of counts, with N_r and N_c columns}
}
\value{
vector of t-a-p parameters with maximum likelihood
}
\description{
create a grid of points in parameter space
}
