% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_model_estimation.R
\name{plot_densities_mcmc}
\alias{plot_densities_mcmc}
\title{plot mcmc draw densities}
\usage{
plot_densities_mcmc(model_draws)
}
\arguments{
\item{model_draws}{Use `fit_counts_mcmc` or `fit_ratings_mcmc` to get the
fitted model, then use the `draws()` method to get the draws, e.g.
`fitted_model$draws()`. Don't forget the parentheses--this is a function
call.}
}
\value{
a ggplot object with densities for each parameter
}
\description{
plot mcmc draw densities
}
