% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{plot_draw_densities}
\alias{plot_draw_densities}
\title{plot mcmc draw densities}
\usage{
plot_draw_densities(draws)
}
\arguments{
\item{draws}{a data frame with one row per draw and one column per parameter}
}
\value{
a ggplot object with densities for each parameter
}
\description{
plot mcmc draw densities
}
