% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{plot_rater_cov}
\alias{plot_rater_cov}
\title{Plot rater correlation histogram}
\usage{
plot_rater_cov(rater_cov, range, bins = 50)
}
\arguments{
\item{rater_cov}{The output from `rater_cov()`}

\item{range}{The output from `rater_cov_quantile()`}

\item{bins}{how many bins for the histogram, defaulting to 50}
}
\value{
a ggplot
}
\description{
Plot rater correlation histogram
}
