% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_model_estimation.R
\name{pull_avg_params}
\alias{pull_avg_params}
\title{Pull average rating parameters as an array}
\usage{
pull_avg_params(params)
}
\arguments{
\item{the}{output of `fit_counts_mcmc()`}
}
\value{
a named array with the average t, a, and p parameters
}
\description{
Pull average rating parameters as an array
}
