% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{pull_rating_params}
\alias{pull_rating_params}
\title{pull the t-a-p parameters from a rating_params data frame}
\usage{
pull_rating_params(rating_params)
}
\arguments{
\item{rating_params}{a data frame with columns subject_id, rating, rater_id,
and at least parameters t, a, and p for each row}
}
\value{
a list with two data frames, one for subject t_i, and another with
rater parameters a_j, and p_j
}
\description{
This function is useful for taking the parameters from a fitted model and
then simulating a rating set from it using `generate_ti_aj_pj_ratings()`
}
