% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{rater_calibration}
\alias{rater_calibration}
\title{Rater Calibration Plot}
\usage{
rater_calibration(rating_params, bins = 10, labels = FALSE)
}
\arguments{
\item{rating_params}{A rating_params dataframe}

\item{bins}{Number of bins to use for the calibration plot, or zero for no binning}

\item{labels}{For bins = 0, plot the rater_id labels instead of points. This might
be useful for identifying outliers. Defaults to FALSE.}
}
\value{
A ggplot object with the calibration plot and error statistics
}
\description{
For each rater compute the fraction of Class 1 ratings and
compare this to the the expected fraction from the t-a-p model parameters
}
\details{
The predicted class 1 probabilities are ta + (1 - a)p, and these
are optionally binned into equal-sized groups. Then the average predicted value and
average observed value are calculated, and plotted. The observed values
are on the x-axis, so it's easy to interpret dots above the 1-1 line as
over-prediction, and dots below the 1-1 line as under-prediction. If no binning
is requested, each rater is a point on the plot, or if labels = TRUE,
the rater_id.
}
