% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{rater_cov}
\alias{rater_cov}
\title{Rater covariance}
\usage{
rater_cov(rating_params, max_raters = 100)
}
\arguments{
\item{rating_params}{A rating_params data frame}

\item{max_raters}{The max number of unique raters to sample, defaulting to 100}
}
\value{
A list with individual `raters` results, including actual and predicted correlation
}
\description{
Rater covariance
}
\details{
This function uses a formula similar to that in the Appendix to estimate the
covariance between ratings based on individual parameters as a comparison point to the
actual correlations of ratings over subjects.

With too many raters, this calculation will bog down. That's why the default
is to sample raters if there are more than 100.
}
