% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{rater_cov_quantile}
\alias{rater_cov_quantile}
\title{quantiles for expected covariance between raters}
\usage{
rater_cov_quantile(rating_params, n_sim = 10, quantiles = c(0.02, 0.98))
}
\arguments{
\item{rating_params}{A rating_params data frame, usually from a fitted model}

\item{n_sim}{How many simulations to run, defaulting to 10}

\item{quantiles}{What quantiles to return, defaulting to c(.02, .98)}
}
\value{
A named vector with the quantiles on the difference between expected
and actual correlation.
}
\description{
quantiles for expected covariance between raters
}
