% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{rater_entropy}
\alias{rater_entropy}
\title{rater entropy}
\usage{
rater_entropy(rater_params)
}
\arguments{
\item{rater_params}{Either a dataframe, typically from
`pull_rating_params(rating_params)$raters` with columns for rater_id, a_j, p_j,
OR the results of `fit_counts(counts)` with the average t-a-p parameters.}
}
\value{
The same dataframe with an additional column s_dot for the
rater entropy.
}
\description{
Calculates rater-specific entropy s-dot, which can be averaged
to estimate the average rater entropy.
}
