% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{rating_calibration}
\alias{rating_calibration}
\title{Rating Calibration Plot}
\usage{
rating_calibration(rating_params, bins = 10)
}
\arguments{
\item{rating_params}{A rating_params dataframe}

\item{bins}{Number of bins to use for the calibration plot}
}
\value{
A ggplot object with the calibration plot and error statistics
}
\description{
Ratings are sorted by pr_c1 and the cumulative sums of
observed vs modeled plotted with comparison to the 1-1 line.
}
\details{
The predicted class 1 probabilities are ta + (1 - a)p, and these
are binned into equal-sized groups. Then the average predicted value and
average observed value are calculated, and plotted. The observed values
are on the x-axis, so it's easy to interpret dots above the 1-1 line as
over-prediction, and dots below the 1-1 line as under-prediction.
}
