% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round while preserving sum}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{vector of numbers}

\item{digits}{number of digits to round to}
}
\value{
vector of rounded numbers
}
\description{
round a vector of numbers while preserving the sum
}
