% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{simulate_tap_fit}
\alias{simulate_tap_fit}
\title{Generate rating samples and model fit}
\usage{
simulate_tap_fit(N_s, N_r, params, n_sim = 100, output = "data")
}
\arguments{
\item{N_s}{number of subjects}

\item{N_r}{average number of raters per subject}

\item{params}{list of t-a-p parameters for the simulation}

\item{n_sim}{number of simulations to run, defaulting to 100}

\item{output}{If "data" (default) a dataframe with the comparison data, or if
"plot" a ggplot object with facet histograms for each parameter
@details
This function uses the solver `fit_counts(counts)` to estimate the t-a-p
parameters, which will combine sampling error with estimation error. In
special circumstances, we can eliminate the estimation error. Use
`simulate_exact_fit()` for that.}
}
\value{
A data frame with the parameter estimates for each simulation or
a histogram summarizing the data for each parameter, including absolute
bias abs(p-t).
}
\description{
Range of accuracy estimates for ideal data
}
