% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{subject_calibration}
\alias{subject_calibration}
\title{Subject Calibration Plot}
\usage{
subject_calibration(rating_params, n_sims = 30)
}
\arguments{
\item{rating_params}{A rating_params dataframe}

\item{n_sims}{Number of simulations to run to estimate the modeled
distribution from the probabilities. Defaults to 30.}
}
\value{
A ggplot object with the calibration plot and error statistics
}
\description{
For each subject compute the fraction of Class 1 ratings and
compare this to the the expected fraction from the t-a-p model parameters
}
\details{
The probabilities for each rating can be averaged to find the *expected* number
of Class 1 ratings per subject, but it's better to get a distribution of
that variation to use in the plot. That's what the n_sim parameter does.
}
