% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{subject_calibration_cat}
\alias{subject_calibration_cat}
\title{Subject Calibration Plot}
\usage{
subject_calibration_cat(cat_ratings, n_bins = 20, n_sims = 30)
}
\arguments{
\item{n_bins}{Number of bins to use in the calibration plot. Defaults to 20.}

\item{n_sims}{Number of simulations to run to estimate the modeled
distribution from the probabilities. Defaults to 30.}

\item{rating_params}{A rating_params dataframe}
}
\value{
A ggplot object with the calibration plot and error statistics
}
\description{
For each subject compute the fraction of Class 1 ratings and
compare this to the the expected fraction from the t-a-p model parameters
}
\details{
The probabilities for each rating can be averaged to find the *expected* number
of Class 1 ratings per subject, but here we want a *distribution* of
that variation to use in the plot. That's what the n_sim parameter does. The
error bars are two standard errors of the proportion estimates for the observed
proportions, with any cases of N < 3 omitted.
}
