% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{sum_list_column_vectors}
\alias{sum_list_column_vectors}
\title{sum component-wise over a given list-vector}
\usage{
sum_list_column_vectors(group_df, col)
}
\arguments{
\item{col}{the name of the column, e.g. t, p, logpi}

\item{df}{a data frame with list-vector `col`}
}
\value{
A single vector that is the componenent-wise (vertical)
sum of the components, collapsing a column of K-vectors into a single
K-vector
}
\description{
sum component-wise over a given list-vector
}
