% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ratings.R
\name{true_params_from_sim}
\alias{true_params_from_sim}
\title{Extract true parameters from a simulation}
\usage{
true_params_from_sim(rating_params)
}
\arguments{
\item{rating_params}{from `generate_ti_aj_pj_ratings()` which must contain
the binary T_i, A_ij, and P_ij values.}
}
\value{
A rating_params data frame with the true t, a, and p values, calculated
from the averages of the T_i, A_ij, and P_ij values.T_i will be binary because
it's the same for each subject.
}
\description{
Extract true parameters from a simulation
}
