% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{vector_line}
\alias{vector_line}
\title{Find points along a line}
\usage{
vector_line(start, stop, steps)
}
\arguments{
\item{start}{vector of starting points}

\item{stop}{vector of ending points}

\item{steps}{number of points to return}
}
\value{
tibble of points along the line
}
\description{
given two t-a-p parameter vectors, return a sequence of points
along the line between them
}
