% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_count_index}
\alias{verify_count_index}
\title{Verify count_index is in the correct format}
\usage{
verify_count_index(count_index)
}
\arguments{
\item{counts}{A data frame with columns N_r, N_c for each subject
There shouldn't be a column n, which indicates multiplicity.}
}
\value{
Zero on success
}
\description{
Verify count_index is in the correct format
}
