% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_counts}
\alias{verify_counts}
\title{Verify counts are in the correct format}
\usage{
verify_counts(counts)
}
\arguments{
\item{counts}{A data frame with columns N_r, N_c, and n}
}
\value{
Zero on success
}
\description{
Verify counts are in the correct format
}
