% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_rating_params}
\alias{verify_rating_params}
\title{Verify rating params are in the correct format}
\usage{
verify_rating_params(rating_params)
}
\arguments{
\item{ratings}{A data frame with columns subject_id, rater_id, and rating}
}
\value{
Zero on success
}
\description{
Verify rating params are in the correct format
}
