% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_counts}
\alias{as_counts}
\title{convert from long form to counts of N_r, N_c pairs}
\usage{
as_counts(ratings)
}
\arguments{
\item{ratings}{A dataframe with subject_id and binary rating}
}
\value{
A data frame with columns N_r, and N_c, and n
}
\description{
convert from long form to counts of N_r, N_c pairs
}
\details{
Unlike `as_count_index`, this function does not associate
counts with subject IDs. Instead it counts up the pairs of (N_r, N_c)
with a column n. This is useful for efficiently fitting the t-a-p model.
}
