% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{bits_per_rating}
\alias{bits_per_rating}
\title{log likelihood by rating}
\usage{
bits_per_rating(rating_params)
}
\arguments{
\item{rating_params}{A data frame with binary subject_id, rating, and a
t, a, and p parameter for the rating.}
}
\value{
The average bits per rating of log likelihood
}
\description{
Given a rating_param data frame from `fit_ratings` or `fit_ratings_mcmc`,
return the average bits per rating of log likelihood using a t_i-a_j-p_j model.
}
