% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{bits_per_rating}
\alias{bits_per_rating}
\title{bits per rating}
\usage{
bits_per_rating(ll_ratings)
}
\arguments{
\item{ll_ratings}{a ratings data frame augmented with C0 and C1, the
likelihood of 0 and 1 ratings. This is the output of `rating_log_likelihood()`}
}
\value{
the average bits per rating over the data set
}
\description{
bits per rating
}
