% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{count_ratings}
\alias{count_ratings}
\title{Count ratings}
\usage{
count_ratings(ratings, inClass)
}
\arguments{
\item{ratings}{A data frame with two columns where the left column is the subject ID
the right column is a rating for that subject.}

\item{inClass}{A vector with the rating values that comprise the in-class. For
example, for a 1-5 scale, you might pick 1:3 to be the in-class, so that
the result is a binary distinction between {1,2,3} and {4,5}.}
}
\value{
A data frame with one row per subject, with the number of raters N_r,
and N_c the  number of raters who rated in the in-class range.
}
\description{
Count the number of raters and the number of raters who rated
in a given range.
}
