% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{e_m_step_cat}
\alias{e_m_step_cat}
\title{Perform an EM step for categorical ratings}
\usage{
e_m_step_cat(cat_ratings, group)
}
\arguments{
\item{cat_ratings}{A cat_ratings object}

\item{group}{If TRUE, groups by rater_id to obtain hierarchical
parameters. Otherwise returns average a, p}
}
\value{
Updated cat_ratings object
}
\description{
Perform an EM step for categorical ratings
}
