% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{exact_accuracy}
\alias{exact_accuracy}
\title{Estimate the accuracy from the count proportions and t, p parameters}
\usage{
exact_accuracy(counts, tp = 0.5)
}
\arguments{
\item{counts}{A data frame with columns N_r, N_c, and n, representing the number
of class 1 ratings. This can be created from `as_counts(ratings)`. The number
of raters N_r must be the same for all rows.}

\item{tp}{the t and p parameters, which are assumed equal (unbiased raters).
This parameter is only required if count_coefs is NULL, since it is used
to calculate those coefficients with `exact_accuracy_coefs()`. It defaults to
.5.}
}
\value{
the accuracy parameter as a scalar
}
\description{
Given counts of class 1 ratings and the t, p parameters, estimate the accuracy
using the closed form. You can either provide the closed-form coefficients or
have it calculate them.
}
\details{
Either tp or count_coefs must be provided. These correspond to the
choices to have this function calculate the coefficients for you, in which
case tp must be specified, or provide the pre-calculated coefficients, e.g. if
you're doing a lot of calculations with them in a simulation. The functions
that start with "exact_" are used for closed-form
calculation of accuracy.
}
