% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{find_solution}
\alias{find_solution}
\title{Fit the t-a-p model to binary ratings}
\usage{
find_solution(counts, init_params)
}
\arguments{
\item{counts}{a data frame with columns N_r and N_c, representing number of
raters and counts of raters who rated class 1.}

\item{init_params}{initial values for t, a, and p}
}
\value{
a data frame with columns for parameters: t, a, p, and log likelihood of the solution
}
\description{
Fit the t-a-p model to binary ratings
}
