% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{fit_counts_cat}
\alias{fit_counts_cat}
\title{Fit categorical t-a-p model (average parameters only)}
\usage{
fit_counts_cat(cat_ratings, max_iter = 20)
}
\arguments{
\item{cat_ratings}{A cat_ratings object (see generate_sample_ratings_cat()).}

\item{max_iter}{Maximum number of EM iterations (default 30).}
}
\value{
A tibble with average accuracy a and guess distribution p.
}
\description{
Fit categorical t-a-p model (average parameters only)
}
