% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap_model_estimation.R
\name{fit_ordinal_tap}
\alias{fit_ordinal_tap}
\title{t-a-p coefficients over an ordinal scale}
\usage{
fit_ordinal_tap(ordinal_ratings)
}
\arguments{
\item{ordinal_ratings}{a data frame in long format with columns subject_id and rating
where the ratings are for a single dimension, with multiple ordinal values,
for example water quality on a scale of 1 = poor, 2 = fair, 3 = good,
4 = excellent. If the ratings are character strings, they will be sorted
alphabetically, so "11" would sort before "2." If the ratings are numeric,
they will be sorted numerically.}
}
\value{
a data frame with columns for each cut point between ratings,
parameters: t, a, p, and ll for the t-a-p model and the Fleiss kappa, as
well as a column to flag degenerate cases, where one of the parameters is
zero or one.
}
\description{
t-a-p coefficients over an ordinal scale
}
