% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ratings.R
\name{generate_exact_ratings}
\alias{generate_exact_ratings}
\title{get ratings that have a given distribution}
\usage{
generate_exact_ratings(N_s, N_r, params = list(t = 0.5, a = 0.7, p = 0.5))
}
\arguments{
\item{params}{t, a, and p, possibly including a0,a1,p0,p1 in a named list}

\item{n_subjects}{number of subjects}

\item{n_raters}{number of raters per subject on average}
}
\value{
a data frame with ratings and parameters, including the input t-a-p averages, the
random effect parameters p_i and a_j and the discrete random variables T_i, P_ij, and A_ij.
}
\description{
get ratings that have a given distribution
}
\details{
This function generates ratings for a set of subjects based on a t-a-p
model. It uses t to determine the 0 or 1 truth values for each subject,
creates random effects p_i and a_j for each subject and rater, respectively, using a beta distribution.
}
