% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ratings.R
\name{generate_sample_ratings}
\alias{generate_sample_ratings}
\title{get ratings by sampling a known distribution}
\usage{
generate_sample_ratings(
  N_s = 100,
  N_r = 5,
  params = list(t = 0.5, a = 0.7, p = 0.5),
  details = FALSE
)
}
\arguments{
\item{N_s}{number of subjects}

\item{N_r}{number of raters per subject on average}

\item{details}{logical defaulting to FALSE. If TRUE, return the full set
of parameters to include the T_i, P_ij, and A_ij values.}
}
\value{
a data frame with ratings and parameters, including the input t-a-p averages, the
random effect parameters p_i and a_j and the discrete random variables T_i, P_ij, and A_ij.
}
\description{
get ratings by sampling a known distribution
}
