% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{generate_sample_ratings_cat}
\alias{generate_sample_ratings_cat}
\title{get ratings by sampling a known distribution}
\usage{
generate_sample_ratings_cat(
  N_s = 100,
  N_r = 5,
  K = 3,
  params = NULL,
  details = FALSE
)
}
\arguments{
\item{N_s}{number of subjects}

\item{N_r}{number of raters per subject}

\item{K}{the number of categories, labeled 1 to K}

\item{params}{A list with t, a, p. t and p are non-negative vectors of length K, which
will be normalized to make probability distributions. If NULL, uniform values are provided.}

\item{details}{logical defaulting to FALSE. If TRUE, return the full set
of parameters to include the T_i, P_ij, and A_ij values.}
}
\value{
A list that comprises the cat_ratings type: data frames ratings, subjects, and raters,
 an integer K > 1 with the number of categories, and a vector labels that
 give names to the categorizes. For the sample ratings,
 values are provided for subject and rater parameters that match the
 generation process and nominal category labels (Class 1, Class 2, etc) are
 created.
}
\description{
get ratings by sampling a known distribution
}
