% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{log_likelihood}
\alias{log_likelihood}
\title{negative log likelihood for binary ratings using t-a-p}
\usage{
log_likelihood(params, counts)
}
\arguments{
\item{params}{parameters t, a, and p in a vector}

\item{counts}{a data frame with columns N_r and N_c, representing number of
raters and counts of raters who rated class 1.}
}
\value{
the log likelihood of the t-a-p model, or 1e6 if the parameters are out of range
}
\description{
Use the implied binomial mixture to calculate the negative log likelihood. It's
negated because the optimizer minimizes rather than maximizes.
}
