% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{negative_log_prob}
\alias{negative_log_prob}
\title{Negative log likelihood of binomial mixture}
\usage{
negative_log_prob(N_r, N_c, t, a, p)
}
\arguments{
\item{N_r}{number of raters as a scalar}

\item{N_c}{number of raters who rated T_i = 1 as a scalar}

\item{t}{mixture parameter in the t-a-p model}

\item{a}{accuracy in the t-a-p model}

\item{p}{guess distribution in the t-a-p model}
}
\value{
negative log likelihood of the binomial mixture as a scalar
}
\description{
given, t-a-p parameters, return -LL for the implied binomial mixture. It's
negated because the optimizer minimizes rather than maximizes. This function works on
scalars because of \code{dbinom}, and is used by the \code{\link{log_likelihood}} function.
}
