% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_fit.R
\name{simulate_exact_fit}
\alias{simulate_exact_fit}
\title{Simulate sampling error with exact formulas}
\usage{
simulate_exact_fit(
  N_s_vals = c(20, 100, 300),
  N_r_vals = c(2, 5, 10),
  a_vals = seq(0, 1, 0.05),
  tp = 0.5,
  use_fleiss = TRUE,
  n_sim = 100,
  output = "plot"
)
}
\arguments{
\item{N_s_vals}{A vector of one or more values for N_s, the number of subjects}

\item{N_r_vals}{A vector of one or more values for N_r, the average number of raters per subject}

\item{a_vals}{A vector of values to use for accuracy, defaults to seq(0,1,.05)}

\item{tp}{The value of t and p, which must be equal. This is used for the
simulations as well as the exact coefficient model when use_fleiss = FALSE.}

\item{use_fleiss}{If TRUE, use the Fleiss kappa to estimate accuracy, otherwise
the exact formula is used. Default is TRUE. There are limitations to the
exact formula.}

\item{n_sim}{number of simulations to run on each combination of N_s and
N_r, defaulting to 100}

\item{output}{If "data" (default) a dataframe with the comparison data, or if
"plot" a ggplot object with facet histograms for each parameter}
}
\value{
A dataframe with the parameter estimates for each simulation or
a facet grid of boxplots along the a axis
}
\description{
Range of accuracy estimates for data simulated with various
accuracies and the other paraters fixed. This function uses the exact
formulas, but is limited to the unbiased rater case (t = p). This function
returns a either the data or a plot.
}
\details{
This function uses the exact formulas for the t-a-p parameters using
one of two methods. The default is to use the Fleiss kappa to calculate accuracy,
which assumes that t=p, but not a particular value. If you choose use_fliess
 it will use the Fleiss kappa. The second method is the linear
combination for a squared as described in Chapter 2: Exact Formulas.
}
