% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_params}
\alias{verify_params}
\title{verify params}
\usage{
verify_params(params, expand = FALSE)
}
\arguments{
\item{params}{A list with the t-a-p parameters}

\item{expand}{If TRUE, the function will expand the parameters to the full
list. This can be convenient for LL calculations to simplify the formula.}
}
\value{
The original params as a data frame if formatted correctly, or an
error message.
}
\description{
Verify that the t-a-p parameters are in the right format.
}
\details{
The params must contain one of the following sets:
- t, a, p
- t, a0, a1, p
- t, a, p0, p1
- t, a0, a1, p0, p1
}
