% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{verify_ratings}
\alias{verify_ratings}
\title{Verify ratings are in the correct format}
\usage{
verify_ratings(ratings)
}
\arguments{
\item{ratings}{A data frame with columns subject_id, rater_id, and rating}
}
\value{
The original data frame if it's right, or an error message
}
\description{
Verify ratings are in the correct format
}
